requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
openai>=1.0.0
python-dotenv==1.0.1
aiohttp==3.9.3
rich==13.7.0
colorama==0.4.6
python-socketio==5.11.1
requests>=2.31.0 